add z4h-fzf-dir-history widget
Zsh for Humans now remembers visited directories and stores them
in a file. The only use of this data is currently via
z4h-fzf-dir-history widget, bound to Alt+R by default. When triggered,
this widget displays the history through fzf. Enter caused the current
directory to be changed to the selected directory. "Repeat" does the
same plus an immediate z4h-cd-down. "Repeat" isn't bound by default
but can be bound the same way as in other widgets:

    zstyle ':z4h:fzf-dir-history' fzf-bindings tab:repeat

Users who have a similar binding for cd-down will likely want it
for fzf-dir-history, too.

The first listed directory in z4h-fzf-dir-history is always the
current directory, so with the bindings listed above pressing Alt+R
followed by Tab is equivalent to Alt+Down. This suggests that maybe
it's a good idea to bind Alt+Down to z4h-fzf-dir-history and stop
using z4h-cd-down.

The maximum size of history can be set with the following style:

    zstyle :z4h:dir-history: max-size 10000

The default is 10000.